validation dependency

+++ src/main/java/github/devluiss/libraryapi/service/AutorService.java
@@ -1,7 +1,9 @@
 package github.devluiss.libraryapi.service;
 
+import github.devluiss.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import github.devluiss.libraryapi.model.Autor;
 import github.devluiss.libraryapi.repository.AutorRepository;
+import github.devluiss.libraryapi.repository.LivroRepository;
 import github.devluiss.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
@@ -16,9 +18,12 @@ public class AutorService {
 
     private final AutorValidator validator;
 
-    public AutorService(AutorRepository repository, AutorValidator validator){
+    private final LivroRepository livroRepository;
+
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository){
         this.repository = repository;
         this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor){
@@ -39,6 +44,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if(possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um Autor que possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -56,5 +64,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    private boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 }
